+++ phpBB2/quick_reply.php
@@ -94,13 +94,7 @@ if ( $total_posts > 0 && isset($postrow[$total_posts - 1]) && !(((!$is_auth['aut
 			'ATTACH_SIGNATURE' => $attach_sig,
 			'NOTIFY_ON_REPLY' => $notify_user)
 		);
-	}else
-	{
-		$template->assign_block_vars('quick_reply.user_logged_out', array());
 	}
-
-
-//	generate_smilies_row();
 
 	$template->assign_vars(array(
 		'U_MORE_SMILIES' => append_sid("quick_reply.$phpEx?mode=smilies"),
@@ -111,56 +105,15 @@ if ( $total_posts > 0 && isset($postrow[$total_posts - 1]) && !(((!$is_auth['aut
 		'L_CANCEL' => $lang['Cancel'],
 		'L_ATTACH_SIGNATURE' => $lang['Attach_signature'], 
 		'L_NOTIFY_ON_REPLY' => $lang['Notify'],
-		'L_NOTIFY_ON_REPLY' => $lang['Notify'],
-		'L_ATTACH_SIGNATURE' => $lang['Attach_signature'],
 		'L_ALL_SMILIES' => $lang['Quick_Reply_smilies'],
 		'L_QUOTE_SELECTED' => $lang['QuoteSelelected'],
 		'L_NO_TEXT_SELECTED' => $lang['QuoteSelelectedEmpty'],
 		'L_EMPTY_MESSAGE' => $lang['Empty_message'],
 		'L_QUOTE_LAST_MESSAGE' => $lang['Quick_quote'],
 		'L_QUICK_REPLY' => $lang['Quick_Reply'],
-		'L_PREVIEW' => $lang['Preview'],
-		'L_SUBMIT' => $lang['Submit'],
 		'L_ADD_SMILIES' => $lang['Quick_add_smilies'])
 );
 }
 $template->assign_var_from_handle('QUICKREPLY_OUTPUT', 'quick_reply_output');
 	
-function generate_smilies_row()
-{
-	global $db, $board_config, $template;
-
-	$max_smilies = 24;
-
-	switch ( SQL_LAYER )
-	{
-		case 'mssql':
-			$sql = 'SELECT TOP ' . $max_smilies . ' min(emoticon) AS emoticon,
-			min(code) AS code, smile_url
-			FROM ' . SMILIES_TABLE . ' 
-			GROUP BY [smile_url]';
-		break;
-
-		default:
-			$sql = 'SELECT emoticon, code, smile_url
-			FROM ' . SMILIES_TABLE . ' 
-			GROUP BY smile_url
-			ORDER BY smilies_id LIMIT ' . $max_smilies;
-		break;
-	}
-	if (!$result = $db->sql_query($sql))
-	{
-		message_die(GENERAL_ERROR, "Couldn't retrieve smilies list", '', __LINE__, __FILE__, $sql);
-	}
-	$smilies_count = $db->sql_numrows($result);
-	$smilies_data = $db->sql_fetchrowset($result);
-	for ($i = 0; $i < $smilies_count; $i++)
-	{
-			$template->assign_block_vars('quick_reply.smilies', array(
-				'CODE' => $smilies_data[$i]['code'],
-				'URL' => $board_config['smilies_path'] . '/' . $smilies_data[$i]['smile_url'],
-				'DESC' => $smilies_data[$i]['emoticon'])
-			);
-	}
-}
 ?>